MAINT: create `_lib._array_api_override` to solve circular import (#23028)
* MAINT: create `_lib._array_api_override` to solve circular import

* lint

* nit

* lint

* Update _array_api_override.py


---------
